Get rid of MSG_NO_ENERGY canned message.
This canned message was used both for when the player is too hungry
to cast a specific spell and for when the player is too hungry to
cast *any* spells. It was confusing in the "any spell" case, because
"You don't have enough energy to cast that spell." suggests that some
specific spell was attempted. When it happened to me I was convinced
that I had some macro attached to 'z' by mistake.

The patch removes MSG_NO_ENERGY and replaces both uses of it with
more specific messages: "You don't have enough energy to cast that
spell." for the case where a specific spell was tried (same as
before) and "You don't have enough energy to cast any spells." for
the case where no specific spell was selected.