Fix an interaction of fallback randarts and acquirement (12277)
I hadn't fully understood one of the ways in which acquired objects
could interfere with unrand generation. After this commit, unrands that
place outside of levelgen (acquirement) or in the abyss or a trove
should always cause later instances of that unrand to simply be replaced
with a fallback randart, and have no other effect on item generation
during levelgen. Before this commit, unrands that were forced by vaults
would be correctly replaced with a fallback if they had already been
acquired, but unrands that were placed randomly (e.g. by | in a vault;
Slime:5 was a good place to check) could be affected by previous
acquirements in a way that had unpredictable results for those items.
I'm pretty sure that this bug had no impact on item generation (or
levelgen more broadly) outside of those specific grids.

This commit may change some seeds as I'm adding a subgenerator the will
impact certain rare cases of unrand placement (mainly for uniques who
can come with a special unrand). These uniques may need further testing.

No save compat for this, so upgrading will lead to the same behavior as
games from before this commit would have.