Fix timing for map view cursor clamping
Because `process_map_command` uses -1,-1 to signal that the cursor
position should be calculated from map bounds, it is wrong to clamp the
cursor to map bounds at the end of this call. (Even worse, when the
command is a level change, it uses the bounds for the current level.)
This fixes the issue with stair destinations not working right with []
in map view.

Resolves #1452

(cherry picked from commit be553b73f1881e9dc65462aeb86542675ae9c9b4)