Don't call redraw_screen while off-level
This call was happening any time the level map was exited, and it is
triggering viewwindow, which triggers show_init, which updates map
knowledge at the player position. A dedicated player could use this to
fully map any visited level by exiting map view on the level to be
mapped while moving around a fully mapped level. This bug was introduced
in e98a5be7b87 which removed the scope protecting the level excursion
object; the immediate fix here is to restore the scope. I'm not sure
that this fix eliminates all ways of causing this...

Resolves #1433