Adjust some catch2 tests
After be553b73, process_map_command no longer clamps the coordinates,
because this function uses a return of -1,-1 to indicate that the
coordinates need to be recalculated. The clamping now happens as part of
the map logic itself. This commit does the clamping explicitly, which is
a bit more trivial but at least exercises known_map_bounds.