Fix a crash for elemental staff fallbacks
This is the flip-side of 64415b183b0; if items roles an enhancer staff,
and upgrades to the elemental staff, it then tried to generate a
lajatang as a fallback. This is the simplest patch for this code path;
it just overrides this and generates an enhancer staff. (Via other code
paths, a lajatang would still place.) These staff/weapon issues suggest
a bigger refactor but I will leave that for later.