Update the Debian dependency list
Python 2 is no longer installed by default on Debian systems, so we
either need to require the python package (and hence Python 2) or
install the Python 3 yaml package. Seems more prudent to do the latter
as Python 2 has reach EOL. Add python3-yaml to the dependencies for the
crawl package.