Remove global crawl_view buffer
There's no need to keep it between frames, except as an optimization,
and I'm not convinced that malloc can't handle an allocation every
viewwindow() call (we're not running at 60fps...). Even if that were
somehow an issue, the correct solution is to pass in a view buffer to
fill out, so that the caller controls the lifetime and visibility,
and then keep that passed-in instance static and private to the file.

The puttext() call in redraw_console_sidebar() can be safely removed,
since that function only needs to redraw the sidebar, not the dungeon
view.