Remove elf bailey
Baileys have a few fun gimmicks. There's a focus on orcs, with a few
related pals (goblins, hobgoblins, gnolls), and some tactical terrain
- letting monsters use polearms and ranged weapons to cause trouble.

Elf Bailey has none of that. Instead, it has a pile of very out-of-depth
enemies in a series of featureless circles. The theme is fun, and you
can imagine a bailey that used elves as 'bosses' (orc-elf alliance!),
but that vault would need to be designed essentially from scratch.

(This may return if Deep Baileys become extant.)