New background: Delver
Shaft traps are fun. What if you had a background that started shafted
all the way to D:5, and needed to find their way back up to survive?

To make it fair, you start with a nice set of consumables: a wand of
digging, ?mapping, ?fear, ?fog, and !haste. In practice, I found that
this was usually enough to avoid sudden centaur- or bee-based death and
to make it back to the top, and that it was a pretty fun and distinctive
start.