Make screen cell's flash_colour property tiles-only
flash_colour is not used for console output at all; instead,
cell->colour is set to the flash colour directly. This commit makes
that clear by conditionally compiling the flash_colour property. The
resulting behavour is the same as before, but the conditionally-compiled
logic in draw_cell is also much more straight-forward.