Remove unnecessary tiles.update_tabs() call
As far as I'm aware, the purpose of this call is to switch to the map
commands tab when the map is opened. viewwindow() already calls this,
however, so it can be safely removed.

Admittedly, if/when viewwindow is refactored to not unnecessarily update
tabs, this feature might break, but that's a small price to pay for
cutting away at the mess that show_map() has turned into.