Remove an ASSERT for unlinked temporary items
This was added in 0de8e755f4 as a diagnostic for hep abyss issues. It
didn't get immediate results, but a crash for this case has finally
shown up:

https://crawl.kelbi.org/crawl/morgue/poncheis/crash-poncheis-20200621-183042.txt

What seems to be happening is that that the player got pulled into the
abyss away from their ancestor, and the ancestor died ("Matia has
departed this plane of existence.") Somehow, at this point, cleanup
didn't happen.

I'm removing the ASSERT for now because apparently once a game gets into
this state it keeps crashing. I *think* that without this check, the
game will work but print unlinked temp item errors.