Display to-hit in monster xv
When examining monsters, display approximate to-hit chances for
your current attack (melee or ranged) next to their EV. Make some
adjustments to to-hit calculation needed to display this EV in an
almost non-duplicated way. Experiments with fsim and matlab suggest
these changes should have minimal impact on play; fingers crossed.

Minor other changes:
- Flatten all UC hit modifiers from slight variance (between +4.5 to
  +6) to +5 for all applicable forms, for simplicity. This may
  have some balance impacts.
- Don't amuse Xom with non-random attacks (which aren't real).

Thanks to elliptic for the advice and WiseOldWolf/SanHolo for the
extensive help and code contributions with this!