Rework corona/halo/umbra to-hit
These were applied after player hit (but not monster hit!) was
rolled. This decreased variance and made to-hit calculation code
more complex for extremely little reason. Instead, move them to
before to-hit is rolled and make halo & umbra apply a constant value,
very roughly the mean of the old roll.