Rename Wizards to Hedge Wizards
Wizards are a weird background. Their name and starting stats both
point towards a 'pure magic' playstyle, but their starting book
is quite poorly suited to such an approach - they aren't
well-equipped to 'blast' foes, and are much better advised taking
advantage of their imps and mephitic clouds to do a little
backstabbing.

This playstyle seems fine and distinctive, so let's make it
clearer to players. Renaming them to 'hedge wizards' shows that
they aren't expected to be masters of 'high magic' (pure blasting),
as does tweaking their stat spread closer to skalds'. (If we want
the extremely high-int spread to be on some background, Cj would
be a reasonable choice.)