Rename Assassin -> Brigand
If you see that you're playing an 'assassin' with a dagger, your
playstyle is very obvious: you sneak around and stab unsuspecting
enemies. But that's not what Crawl's Assassin is!

If you want to stab monsters after the first few floors, you want
to be playing an Enchanter, Earth Elementalist, or someone else with
the ability to incapacitate foes. Crawl's stealth mechanic isn't
meant to support 'purely mundane' stabbing, but the Assassin's very
powerful theming is extremely effective at convincing new players
otherwise, and then frustrating them when it doesn't work.

A Brigand ia more brutish figure, an outlaw who leaps from the woods
with an axe and chops you up. They're not sneaky in the same way that
an Assassin is, and so, I hope, won't encourage the same confusion.

Other considered names: 'ruffian', 'thug', 'skirmisher', 'brigand'.