Fix bug related to finger amulet when hydra transforms.
Recently, if hydra transforms to another form equipping finger amulet with ring, then they are not well melded. It can occur the bug.
Only the ring is melded, but your amulet doesn't. it means that you can remove your amulet without considering the ring.
By this fix, you can't remove the ring on your finger amulet when the ring is melded.