Render dungeon zaps onto vbuf
This commit merges the rendering of zaps into the vbuf. Previously,
these were rendered separately via add_overlay() and clear_overlays() on
tiles, and blitted directly to the output buffer on console.

That approach had a number of issues. First, resizing the local tiles
window when zaps were shown caused them to vanish; this was easy to
reproduce when resizing the window while unlocking the entrance to Zot.
It involved global state that had to be kept in sync with vbuf updates.

It was also a poor fit for the game semantics; zaps are caused by
in-game elements that are part of the game simulation, so they should be
rendered in the vbuf, unlike objects like cursors that are part of the
UI.

This commit has a number of parts. First, the calls to draw zaps were
replaced with new interfaces that are implemented in view.cc. There is
still some static state here, unfortunately, but this allows zaps to at
least be rendered to the vbuf. Next, zaps are rendered into the vbuf, at
the bottom of draw_cell().

Actually drawing this vbuf requires minor modifications to rendering
code: on local tiles, overlays (and all other tiles for each cell) are
sorted by virtue of the fact that the local tiles renderer must pack
each overlay tile into separate per-texture buffers. Webtiles has no
such restriction, and so we must render overlays from different textures
in two steps to get the desired effect (zaps on top of everything).