Crash when trying to save the game during a level excursion
Doing this is really, really bad because you.pos is affected by the
excursion (among other things). However, this is a standalone commit in
case it needs to be quickly reverted for some reason I hadn't
anticipated. (An alternative would be to just returnin this case.)

(cherry picked from commit 93d59993c006efd640d0d788809080720d221272)