Add some sanity checks for flashes/more
Right now, if a flash happens during a level excursion, it can trigger a
screen write out of bounds error. This is a low-level check to prevent
this from ever happening, though I think the bug is probably in how
viewwindow is called during map view excursions. (To replicate, set a
message flash color for `warning` and write an annotation with `!` while
off-level in map view.)