Fixup in case a bug leaves a player in a wall
Obviously this shouldn't happen in the first place, but when it does,
the game is pretty much stuck without manual save file repair (needing
both admin and dev work) because of an ASSERT in movement.cc. This
commit fixes up such games on load, by moving them preferentially to
seen stairs. (The reasoning is that when this happens, the crash is
likely to have put the player at an arbitrary position).  By doing this
on load, we should still get crashes that lead to this.

This is specifically aimed at this game:
    https://crawl.xtahua.com/crawl/morgue/sentinel/crash-sentinel-20200627-151113.txt
where a crash during a level excursion stuck the player kind of randomly
(based on actual pos) on the level being viewed. The crash and setup for
this are addressed in other commits. I'm doing this generally because I
have manually fixed several games that get into this state somehow or
other in the past.

(cherry picked from commit 2bcdb004479263eae0a2acee5312cce703015e0c)