Optimizations for drawing many gold piles
These optimizations hit points that are the heaviest special cases (as
revealed by profiling) for gozag characters in situations with lots of
enemies, e.g. megazigging.  They lead to measurable improvements but
things are still not exactly fast.

This commit is treating the symptoms, not the problem; the problem is
that viewwindow is called many, many times and each time it is called
rebuilds just about everything that is displayed, leading to some kind
of fairly bad polynomial penalty. For example, on a test gozag char that
is on a megazig 13, with ~130 items in view and only about 10 cells
stepped on, `stash_annotate_item` alone gets called 145000 times via
several different code paths when doing a full autoexplore to pick up
items; when just standing still and hitting esc there are still hundreds
to thousands of viewwindow calls. (So the reason special casing gold in
that function helps is because a big proportion of these items are gold
piles.) The real solution therefore should be to figure out how to not
rebuild this information on every single viewwindow call, and/or to call
viewwindow less. One starting point might be moving the show_init call
elsewhere (this entirely tears down and rebuilds map knowledge every
time). However, getting such a change right across all build targets is
not easy.

This commit may look like yak shaving, since gozag megazigging is not
extremely common, but the reason I'm focusing on it is that when someone
is megazigging on a server, the resulting cpu impact is quite noticeable
across everyone else currently playing to the point of causing mass
timeouts occasionally on a busy server. In my testing, which is not
representative of server scenarios, there is a 20% absolute speedup with
this commit (on a debug build) but cpu usage also suggests that the
resulting rendering code paths become more io-bound, so hopefully this
will lead to better sharing.