Move a bunch of preprocessor macros out of AppHdr.h
Currently, most headers are not self-contained; they often refer to
symbols that they do not themselves import. This is pretty awful and
tends to flummox code analysis tools, as well as making automated
refactoring quite challenging (since you cannot simply include a
header).

This commit splits a bunch of widely used macros out of AppHdr.h, which
is the pre-compiled header for DCSS. Unfortunately it's accumulated a
bunch of other cruft which really doesn't belong there; these macros are
commonly used in headers as well, but it's unwise for headers to import
AppHdr.h, since it has many dependencies by design.

We're keeping the 'using namespace std' for now, but it'd really be
better if we can slowly remove this... it's especially bad to include in
headers, since it extends to anything that includes them.