0.25-specific seed fix for heliophobic_arrival_battle_scene
This vault had two issues: (i) in the vault code, the number of rng
calls depended on player species (so could affect the seed differently
depending on whether the player is a corpse species). (ii) the corpse
generation code actually generates, places, and kills a monster, which
does all sorts of stuff; but the main impact is that it will along the
way generate weapons depending on species, which leads to substantial
differences in rng rolls per species.

In trunk I have fixed this in a general way; see 644bcb52. However, I
noticed that Yermak's recent turncount record during the tournament
(seed 10834663189915259185, see [1]) places this vault, and I wanted to
allow players to try out replaying this seed in the 0.25 branch (without
having to play 0.25.0). I have therefore chosen to stabilize seeds with
this arrival vault to always match the DD start in 0.25.0. (I looked for
more elegant ways of accomplishing the goal, but because of the
complexities of how corpses are placed this is the best I came up with.)
This will obviously affect other seeds as well, but 0.25.0 is still
available if people do want to go back and replay other species variants
of seeds with this vault.

[1] https://crawl.kelbi.org/crawl/morgue/Yermak/morgue-Yermak-20200628-195118.txt