Give Artificers a choice of starting weapon
Artificers have neither an in-game reason to start with a particualr weapon (UC for Tm, dagger for En) nor a  thematic reason (dagger for Assassin).  The short sword they start with in 0.25 is an example of WCHOICE_PLAIN, so give that to Ar and take the explicit short sword from their starting inventory.