Fix epilogue reloading
I still don't fully understand this bug, but under some circumstances,
despite the changes to caching in 50bb873abc7, vdefs contained mapdefs
that did not have the epilogue. (The easiest way to reproduce was to
save after a level with an epilogue is generated but before entiring,
fully quit crawl, reload the game, and enter the level. Restarting to
the main menu after saving wasn't enough.) This fixes the issue by fully
loading the vault to fill in the epilogue, which seems non-ideal, but I
infer that it may be expected from other cases where load() is called.
It may be worth revisiting just serializing the epilogue...

This was leading to a lack of milestones/entry announcements for wizlabs
when the epilogue reload failed.

(cherry picked from commit 4a859bdc29e1467350147016b25e0ca2efb97600)