Use last game type for startup menu, cleanup
This is aimed at fixing a specific bug: with `restart_after_save`, if
you start a new game and then save, the name is (unexpectedly) cleared.
This is because of the static variable used here. However, this change
generally does a better job of selecting the default main menu for a
number of restart cases. The one thing I'm not sure about is that if a
sprint or tutorial game ends, it returns you to the submenu instead of
the main menu.

This commit also refactors a lingering static variable as a member
variable.

(cherry picked from commit dbaac098b9f9a9b6b9b197f409c2de90fdae96fc)