Use a smarter cprintf in output.cc
cprintf is not region aware, and so can easily leave the cursor outside
of the region it is writing to. One way it can do this is because if it
prints right up to the edge of the screen, it will wrap to the very
beginning of the next line. If the starting region is the mlist, this is
guaranteed to happen and leave the cursor in an invalid position. This
commit switches out cprintf for a region-aware version that will wrap
relative to the current region. (This existed but wasn't widely used and
needed some sprucing up.)

This commit also adds a bunch more ASSERTs, and minorly fixes up some of
the monster pane logic.