Remove the Aura of Abjuration and Recall spells
These are the only summoning spells that don't actually summon something,
and neither is especially useful. Killing the summoner (or teleporting
away) has been the best way of abjuring things for some time, and recall
is already built in to the gods that make it relevant with
long-duration/permanent allies.