Fix crash when playing hints mode on local tiles
This was crashing because the tutorial cursor tile is not in the main
tilesheet. The call to pack_cursor() above already handles both checks,
and adds the tile to the correct buffer, so this code can just be
removed.

I suspect the reason this didn't crash before was because previously the
tile ranges overlapped, and so a completely different tile was getting
added to the main buffer.