Remove the 'known' trap designation
Traps are always known to the player, and the 'known' term was supported
as a no-op since so many vaults still used it. This commit removes
parsing of it and removes all occurrences of it DES/lua and the
documentation. Additionally clean up some trap code that referred to
traps being known.