Render direction chooser rays onto vbuf
This commit replaces direct blitting to the screen via putwch on
console, and the use of an additional tiles buffer on tiles, with the
new console and tile APIs for adding overlay tiles to the dungeon view.

This isn't as good a semantic fit as it was for zaps; those are clearly
a part of the game simulation state, whereas this is reusing the same
overlay buffer just to simplify rendering. And now with that in mind,
perhaps zaps should be rendered by having a list of explosions and other
environmental effects in env?

Anyway, the result of this commit is that almost all rendering of the
dungeon view now goes through the vbuf, and that uses of the ugly
is_tiles hack for viewwindow is now almost entirely removed; slowly but
surely, the entangling of tiles and console rendering is being undone.