Have traps require LOS_NO_TRANS
A number of vaults put monsters and traps behind glass, but the vast
majority of these only intend the traps to be relevent when the player
enters the vault. Traps triggered by monsters in these vaults when the
player is safe on the other side of glass generate a lot of message
spam. Hence this commit makes monster trap activation require
LOS_NO_TRANS to reduce this spam.

A couple of vaults do depend on monster trap usage behind glass, and
these will be reworked in a subsequent commit.