Remove a dubious string::reserve call in websockets
In my profiling this is doing *way* more harm than good in the context
of _send_map and _send_everything, and a very large map send drops from
4-500ms to under 100ms when I remove it. This is obviously very
dependent on stdlib implementation but my guess is that trying to
micromanage memory allocation in fairly small increments like this is
always going to be a bad idea. It's possible that a larger reserve call
would do some good before sending the whole map, or before
_send_everything.

Also, a minor simplification for appending a single character.