Simplify eating related code and interface
Now that only ghouls eat chunks, a lot of internal complicated code
regarding chunks and deciding when and why to eat them can be deleted.
Good god food conducts also don't make sense, so they're removed.

The restriction on cannibalism remains for hydra form for flavour
reasons, but could be removed for further simplification as it isn't a
significant mechanical impact (there's no way to violate it now and
incur penance).

This commit also makes the "eat" menu a thing of the past, press e to
eat if food is available, no prompt or consideration, a ration on the
floor will be eaten first. Ghouls still have rations available, but a
subsequent commit will pin their rotting rate (instead of making it
hunger dependent) and prevent them from eating rations.

The option `auto_eat_chunks` is replaced with `auto_eat` which does what
it promises: automatically eating if it makes sense to do so (Hungry (or
rotted), not in danger, food available).

Some butchery and eating options remain, a subsequent commit will clean
these up.