Uncouple Ghoul rotting from hunger state
And prevent Ghouls from eating rations. A further food simplification;
Ghouls have their own "clock", the rot clock, controlled by chunks.
Coupling the opaque hunger state clock to the speed of the rot clock
was an extra layer of complication and mostly not noticable for ghouls:
by the time a Ghoul was in a low corpse area they could have a fat stack
of rations and remain above the threshold. If the clock on Ghouls is too
loose, the rot clock can be accelerated.