Remove hungry ghosts
They were never a good monster, just sort of a funny joke. (See
some of the vaults that I'm removing and tweaking here.) Now that
hunger is startlingly real, hungry ghosts are just a vicious, unfun
trick for anyone dumb enough to melee them, in the same category as
the jellies and pulsating lumps of old. There's no redeemable core
to them, so let's bid them adieu.