Previously, phial of floods did three things:
a) It fired a Primal Wave at the target
b) It created a pool around the target
c) It waterlogged monsters in the pool

Primal Wave deals damage, potentially knocks back a tile, and creates its own pool. This led to phial of floods creating two different overlapping temporary pools, one waterlogged and the other not.

This commit consolidates everything into Primal Wave. Phial no longer manually creates its own pool; instead player-created Primal Waves now scale with power like the Phial pool did, and apply waterlogged appropriately.

This commit also reworks Primal Wave's pool generation. Previously, it created the center tile manually, and then randomly added tiles on a 2-radius circleLOS ring to it. I switched it to use the Phial pool generation method which uses a fair distance_iterator to fill the pool from the center. The results are fairly similar without the weird spoilery effects of circle-LoS math.

This commit also incidentally solves the issue where the actual target of the phial could be pushed out of the pool by the primal wave and miss getting waterlogged. Now that we silence based on the Primal Wave pool (which was always centered on the pushed-back monster) you're guaranteed to at least waterlog your target (if it's vulnerable.)