Inject custom rendering into viewwindow()
A longstanding source of rendering bugs in crawl is the accretion of
various bits of global state used to draw various parts of the main
UI view, all interacted with in various ways both from viewwindow
itself and from the parts of crawl that alter rendering (e.g. rendering
zaps, highlighting summoners, drawing beams, etc.). Good examples of
this are anything that uses view_add_tile_overlay in conjunction with
viewwindow().

The result is that each part of crawl that does any rendering at all is
extremely brittle, since it must depend on this complex global state,
and has to know what the current state of that global state is. This
makes each piece of rendering code work fine most of the time, but
inevitably break when used from a different context which violates these
assumptions. It also makes some things difficult to implement or very
inefficient because multiple viewwindow calls are required.

One of the worst offenders: directn. The direction chooser's drawing
code draws beams and highlights summoners. It used to do this via a
complicated series of calls to viewwindow() that was required in order
to add tile overlays, then writing directly to the screen to overwrite
the console output with the required alterations.

The new code works by allowing the injection of custom render behaviour
into the viewwindow() call. This allows any kind of custom rendering,
instead of limited functions hard-coded directly into viewwindow(),
while at the same time limiting the scope of these modifications; there
is no global scope. You also don't need to understand the viewport
rendering system; simply implement a type and provide an instance, and
viewwindow() will call its callbacks when needed. You don't need to add
tiles to a global buffer or blit directly to the screen; just modify the
provided crawl_view_buffer.

A nice side effect of this is the near eradication of the tiles_only
parameter to viewwindow(). The only remaining use is in tileweb.cc,
where it is required to prevent scribbling over the console map display
while panning the tiles vbuf. One possible approach I am considering is
to inject the rendering behaviour of the base map itself, but that's a
change for another commit.