Remove ability/evoke+hunger interactions
Reasons: (i) hunger is mainly used as a secondary cost for things that
already have some cost or other gating that is actually impactful on a
tactical level (MP, cooldowns, piety) (ii) the costs assigned to
abilities were really all over the place (so consider this balancing
them to 0), and (iii) given the absence of spell cost the relatively
idiosyncratic choices here are not very intuitive at all and quite hard
to keep track of.

I guess the most controversial change here would be berserk, which
doesn't have an initial cost at all now.  On the other hand, it already
has a cooldown, passing out effects, etc. that already provide gating
and a *lot* of tactical balancing tools if tweaking is needed. If it
needs strategic balancing, I would recommend piety (perhaps
heroism-level or lower), but it's not clear to me if it really needs
this.

The one remaining food cost I have left is for staff channeling (energy
/ wucad mu). I'm not sure it's ideal that this stay either, but I don't
have enough of a sense for how channeling is currently balanced (if it
is) in order to adjust. Sif channeling is consistent with other god
abilities.