Rework the Sword of Power
Sword of Power is fun and evocative, but a little underwhelming in terms
of power level. Increase the enchantment from hp/10 to hp/7, and remove
the cap of 27 for simplicity.

Without any swap penalty, players are incentivised to carry their next
best long blade as a swap when the damage falls too much. The Sword of
Power requires commitment from the player, and so gains *Contam.