Remove rations
Motivations: see many, many years of discussion, experiments in forks,
etc. A discussion in ##crawl-dev on 7/7/20 revealed that nearly all
active devs were on board with something like this step as a followup to
de-chunking (plus a surprising amount of agreement on other long-term
clocks we might try, though this commit does not address any alternative
clocks)

This is a proof of concept commit at this point, and there are probably
a bunch of things that could use further tweaking (especially the vault
modifications, where my des skills are not always up to the task). For
non-carnivorous species (which is to say, non-ghouls), they are marked
as foodless. Ghouls stay as-is, though of course no longer get access to
rations. This may need some tuning as well, since on a few ghouls I
checked they did eat rations from time to time as a backup -- since now
only ghouls eat chunks, the obvious knob is nutrition per chunk. I only
deleted a few vaults altogether, and there could be ways of restoring
some of these.

In the future it may be worth considering how ghouls could be cleaned up
-- most of the rather substantial food infrastructure remains in support
of them.