minislime_mu tweaks
- Replace the ration with gold. It was originally a royal jelly,
which was a pretty good joke, but the ration isn't a good joke and
we don't have a royal jelly ration tile. Gold is at least a good ref
to TRJ's description.
- Despoiler the vault's walls. They've been spoilery since they were
added (766f9b2f52f45f, over 10 years ago) about 3/4ths of the time and
there's really no reason for them to be!