Convert Ghouls to heal-on-kill
Ghouls' chunk healing gimmick was very powerful and very flavourful.
This commit attempts to preserve that flavour while paving the way for
removing all OBJ_FOOD class objects.

Ghouls now heal from killing the living. The exact numbers are a reduced
version of the old healing from eating a single chunk, provided on a
coinflip when the monster is killed.

In discussion with elliptic there was some concern that passive heal on
kills (without the turn penalty) could be too strong. These numbers are
by no means final, and further tuning to Ghouls may be needed (apts &c.).

There was a further concern that these new Ghouls would feel too much
like Vine Stalkers; since the relationship to regeneration, potion
healing, and MP are significantly different I don't think that will be
the case in practice.