Replace Artificer starting Short Blade skill with Armour
Response to feedback on PR#1470.

Upgrading to WEAPON_PLAIN was seen as too much of a buff, but the current Ar not signalling that it is a melee background was noted.  This PR does not change the starting short sword item, but changes SBl skill to Arm.  Artificers now also get the +1 to Arm or Ddg based on aptitudes that AK and CK get.