Replace Artificer starting Short Blade skill with Armour
Artificers have neither an in-game reason to start with a particualr
weapon (UC for Tm, dagger for En) nor a  thematic reason (dagger for
Assassin). Starting with Short Blades skill presents an opportunity cost
for an artificer looking to use what the dungeon has given, and steers a
newer player away from this all important practice.

This commit moves the starting shortblades skill to defenses. Artificers
now start with Armour 2/Dodging 1 or Armour 1/Dodging 2 based on
aptitudes, like AK and CK.

Everything old is new again, so this returns Ar to starting with a club
(a weapon notable mostly because one wants to switch away from it)
and changes the short description to better communicate to the player
that Ar is "starting with wands" not "set-up for a wand game".

[ Committer's note: Closes #1484. Squashed, rebased, reworded the commit
  message, and fixed whitespace. ]