Remove chunks and all OBJ_FOOD
With Ghouls transitioned away from chunk eating to heal on kills, the
last food item type can be removed. A few vaults lose a bit of flavour,
but for the most part are preserved with corpses or KPROP: bloody. One,
the food pyramid, just didn't make sense anymore.

There may still be some code clean-up to do after this commit.

Chunks live on in max_monster_chunks, used for computing knockback
effects, simulacra, and Hydra form vore heal. Perhaps in the fullness of
time this can be cleaned up, but as its internal and not player facing
it is less urgent.