Charm dispersal 🔕: Move Ozocubu's armour to ice only
Ozo's Armour has a good tradeoff: you get the bonus as long as you don't
move. This is fine as an ice only spell, like ramparts.

This commit also changes Ozo's to work in all armour with an ER based
penalty to its bonus. Currently the penalty is ER/2, meaning that in
crystal plate max power Ozo's grants only 2 AC.
Training Ozo's to maximum power in CPA is sub-optimal outside of
megazigs where everything is available; even then 2 AC that can be kept
only when not moving is sufficiently marginal that the spell shouldn't
be a no-brainer. If this is wrong the ER restriction can be restored.

If Ozo's armour is too powerful at L3 single school the level can be
bumped.