Charm dispersal 🔕: Move Excruciating Wounds to Necromancy only
Excruciating Wounds was on the borderline. The noise made it a bit
impractical to cast before every fight, but with sufficient spellpower
stairs and retreating to cleared areas could tediously be used to get
the effect for minimal cost.

This changes EW to be single-school, as well as reduces the casting
noise. In exchange, EW prevents unwielding the weapon for the duration
and applies a wizardry penalty (currently the same as spider form). The
reduced noise reduces the incentive for retreat-casting and the
inability to voluntarily cancel the wizardry penalty makes the cost for
use ongoing instead of one-time, which is a more interesting decision.