Remove Sticks to Snakes
This spell was overpowered and broken beyond redemption. Its continued
existence was the result of flavour nostalgia and a lack of initiative,
but uncapped unlimited non-abjurable allies for 2mp that swarm and
constrict really does not fit in modern crawl. It's time to say farewell.

The previous commit reflavoured Song of Slaying as a transmutation, so
Wereblood replaces S2S in the transmuter starting book, giving the book
a complete focus on melee enhancing spells.