Rework the Orb's translocation interference
The translocation restrictions in Zot are very spoilery, and with
controlled blink no longer existing as a spell it should be fine to allow
using scrolls of blinking to work as normal instead of degrading them to
uncontrolled. They're supposed to be a powerful limited resource, so
making them suddenly near-useless at the end of the game doesn't feel
great, and they're also more valuable used early to prevent death when
the character is weak. If hoarding them until Zot turns out to be a
problem, the generation rate could be adjusted.

In addition, don't disable Passage of Golubria, but instead increase the
fuzzing of the portal placement substantially and make it time out more
quickly.