Charm dispersal 🔕: Remove the Charms school, skill, and Skalds
With the remaining charms spells moved out of the school, there isn't
anything for the skill to be useful for. In designing the forthcoming
charm-effect-items it became clear that for the most part they will not
scale with skill; for those that do, we already have a
magic-item-interaction skill: Evocations. In addition to removing the
skill, the school is removed (it was still used for monster spells to
tag them as buffs, but all true buffs are also flagged with selfench)
from spschool type and miscasts. (Some of the miscast flavour lines are
preserved in Hexes miscasts.) This also means removing the Book of
Battle (now mostly empty) and Charm randbooks.

Removing Charms as a school of magic leaves the Skald start without its
core identity: using Charms magic to enhance melee. From a fundamental
design perspective, there are already several backgrounds that much
better embody the use-magic-to-melee better: Warper (positioning),
Transmuter (forms), Enchanter (debuffs), Hedge Wizard (magic grab bag).
Without a separate magic school for their identity, Skalds don't have
much in the way of a background concept. It is possible that the
resulting charm items might make for an interesting adventurer starting
kit of some kind, but that is probably better done by adding them to a
chance in a wanderer start.