Charming item 🔔: Scarf of Shadows
Re-implement the Darkness effect as a scarf ego. This ego also currently
gives 1 spell de-enhancer, since reduced LOS is very strong and allows
more spells to be in range at the edge of LOS. If that makes this item
too weak it can be removed.

The scarf slot was chosen for this effect because it seems like a better
fit compared to the amulet slot, mostly flavor wise. The effect could
work equally well as an amulet, but it seems like a better tradeoff with
the other cloak alternatives than the other amulet alternatives.

Shadow ego icon by CanOfWorms.