Start Artificers with a spectral club
The spectral ego uses evocations, which keeps with the theme of an
"evocable" start. Without weapon skill the club is still hard to use,
and the boost from the brand is limited. Even with evocations and weapon
training a spectral club quickly falls off.

From a gameplay point of view, this should give Ar some skill-based
survivability tools that aren't permanent consumables.

From a flavour stance this is really cool.

From a new-player experience I worry that it will mislead new players
into sticking with Maces & Flails; the hope is the starting description
and skill training will be sufficient to steer past the flashiness.