Fixup library spells unconditionally (hitemv+hard)
This code was originally added specifically for the positional magic
changes, but 61946d8 used it as a more general removed spell fixup to
handle confuse. However, the library calls in particular were
conditioned on an earlier minor tag (from 688896fa), whereas the
memorized spell calls were not -- and so the fixup never got called for
games started after the original minor tag, and confuse could still
exist in someone's spell library.

There's minimal cost to not conditioning on the minor tag, so move this
more towards generic removed spell code than positional-specific code.