Revise Beastly Appendange
Beastly Appendage has two issues at present.

The first is a charms-like "always cast this low level spell for
marginal benefit", the condition isn't quite always, but any melee
character without boots or a helmet who isn't using other forms should
use this (as should any Op).

The second is that the possibilities of Horns and Talons are not
balanced against eachother, so the player is incentivized to re-cast
until the better aux is granted.

This commit addresses both of those, while buffing the spell slightly to
address concerns that Tm is too weak without Sticks to Snakes as a
crutch. Beastly appendage now melds aux armour slots, weapons, and
shields and gives Horns 2 and Talons 2 (Op keep tentacle spike 3). Two
level 2 aux attacks was selected over a single level 3 so that the spell
synergizes better with Wereblood.