Exclude load_games.py from lint
Sort of too bad because this is one of the few files that was passing,
but something weird is happening with flake8-isort -- it seems to be
producing an error for a missing blank line, but when I add the blank
line running `isort` removes that line. Some day someone should do a
real lint pass sorting out this stuff but it is not me, today. (In this
case the solution may be not to use flake8-isort?)

(cherry picked from commit e2f3d4a7ca955ff130fda8b6440f5ff436b3dbdb)