Simplify Battlesphere targeting and spell schools
This commit simplifies the targeting & triggering of Battlesphere to
make it work better with positional magic. The battlesphere is still
synced to the player's conjurations, but now picks a random target that
the caster can see to fire at and is triggered only by conjurations
school spells (except Battlesphere and Servitor).