Allow charge to bypass terrain effects on intermediate squares
Processing the moveto effects for intermediate spaces creates a lot of
technical complexity as well as a lot of special cases that are
difficult to communicate to the player unless we go with all (which is a
prompt/check/code headache) or only the destination (clear, single
prompt for target location, opens up new tactics): will noxious bog or
clouds in the interim effect me? What about traps? Which ones, why and
how?

This commit goes with the second option: only apply the destination
terrain effects. The safety from danger is mentioned explicitly in the
ability description, and flavoured as rolling fast enough to avoid the
dangerous effects of terrain. The destination effects may include
getting webbed or netted; in this case the attack is aborted (so roll
into webs or net traps at your own risk).