Reimplement direction chooser UI as a widget
Now that the direction_chooser uses a view_renderer to inject its custom
rendering code into viewwindow, wrapping this stuff with a widget is
more or less straightforward.

Some complications: the UI framework itself will first draw all current
tilesdl regions before drawing any widgets, since this is the first
widget that is not a popup to be drawn over these regions. The direction
chooser must render these itself, however, as it calls viewwindow with
custom rendering commands. To prevent them being called twice, we add
the ugly ui::should_render_current_regions hack. This can be removed
once the code drawing the current regions is extracted to a 'main view'
widget; this requires extensive modifications to a lot of code in order
to pull all game input through the UI framework's event loop, since
otherwise the main view widget wouldn't be drawn, and *that* can't be
done until directn (and viewmap) are wrapped in widgets. The long and
short of it is that this hack is required to break the catch-22 cycle.

Only part of this needs to be conditional; the maybe_redraw_screen() is
always needed, in case of resizes.

Another less major complication is that we must pump at least the
handle_mouse part of the tilesdl main loop, in order to update the mouse
cursor position on the map region, which requires making handle_mouse
public.