Remove redundant redraws in direction chooser
Now that the direction chooser is implemented as a widget, it doesn't
need to call update_screen() itself; the ui framework will do that once
everything in the widget tree has drawn to the screen.

The range_view_annotator object does not need to call viewwindow, since
the direction chooser widget will immediately do an initial render; so
it can be reduced to just flipping some state.

We also don't need to clear the screen after describing tiles, since the
render handler will just redraw everything.