Track time taken per level
This will only apply to new games after this commit unfortunately, since
it's not info we were tracking already. This implementation is pretty
quick/hacky, but it would be easy to simply drop the prop + code
altogether once we no longer need this data (said everyone who
introduced a new player prop, ever).