Fix excessive redraw when autofiring ranged weapons
This issue seems to have been caused purely by excessive redrawing
of the screen, even though the content being drawn was exactly the
same; terminals naturally lack a way to express any sort of framing or
do atomic updates. This was caused by autofight.lua's implementation
opening the targeter, issuing several commands to move the cursor, and
then firing; each command caused a re-render.

This wasn't an issue on local tiles or non-web console, because they
won't re-render while there's still input. Unfortunately that's not yet
possible on webtiles; I believe there are still one or two UIs that
require a render call in order to get their text painted and sent to
webtiles.

The way around this I've used here is to split buffer swapping /
updating out of rendering, and call it independently. Webtiles updates
are managed separately, so calling render() but not swap_buffers() will
still result in correct display on webtiles. The practical upshot is
that rendering is now skipped on webtiles console if there are pending
key events, which fixes the excessive redraw issue with autofight.lua

Fixes #1492.