Don't give Artificers branded starting weapons
Giving them a weapon with a brand that scales with evocations suggests
that they're supposed to continue to focus on that skill primarily (even
though it's on a very weak base type of weapon), but other recent changes
have been intended to make it clearer that the starting weapon is
something that's supposed to be abandoned and they should adapt to their
finds - the starting wands help survival but there's no real support for
a pure evocations playstyle.

If starting with a spectral weapon is really desired, they should most
likely be somehow rebalanced to get a standard weapon choice and some
starting weapon skill.