Reveal pile contents in los
Requiring players to walk to multi-item piles is a simulationist relic
that is mostly an interface screw; with monster items identified and
necromancy "corpse sense" for corpse using spells, the player can tell
what is in a pile (more or less) in advance, but must navigate their
character to the cell to confirm that information.

This requirement prevented 'x' from describing piles effectively, since
describing a stash has (mostly vestigal) requirements for communicating
the conditional nature of the knowledge.

With this commit piles are no longer obscure things you can only see the
top of. If an item is in LOS, it can be viewed with all of the usual
view commands.