Improve validation of spflag::monster
This was very inconsistently set, because preivously an error would be
generated only for player spells with a monster flag, but not for
non-player spells without a monster flag. To do this required a tiny bit
of extra work to check wand spells and a few idiosyncratic cases
(resulting in yet another static const unordered_set unfortunately), but
found a whole raft of cases where the flag was missing.

This might be better as a catch2 test (along with some of the other
things validated when initializing spell data), but I don't yet
understand how to deal with cases like this which involve multiple .cc
files in catch2 tests.