Replace summon swarm with shadow creatures in fedhas sprint
The implementation for summon swarm was removed in d48e5fa4009, but
fedhas sprint places sentient oklobs that can cast spells, including
this one. Since summon swarm was not actually marked as removed, this
didn't show up as an error until one of my previous commits improving
validation of removed spells etc. This commit uses shadow creatures
instead, and thanks to a tip from kate-, sets the monster list for the
sprint to match that of the old summon swarm spell.