Don't put D:1 plants to sleep (purplemustard)
Recent fixes to D:1 monstergen that again generate all monsters with
BEH_SLEEP, combined with the butterfly fix in 6c26fa703df9 led to any
D:1 plants appearing to be asleep. Since BEH_SLEEP is meaningless on
firewood, simply skip setting sleep-related properties when building
monster-info for firewood, and consequently don't show it to the player.
(So technically, they do still count as asleep() but there's no reason
to make that visible.)