Simplify vehumet + library interaction (11636 etc)
In the days of books, if you memorised a veh offering and then forgot
it, you would be out of luck without that book on hand. This behavior
was more or less preserved into the new world of libraries. However, now
vehumet offerings are the only way in the game to have memorised a spell
that is not in your library, and therefore, a regular source of
confusion. This change adds a veh gift to your library if you choose to
memorise it, removing this (currently) odd exception entirely.

I've also removed a bunch of bools that were only used for the save
compat case, from back when upgrading from games with inventory books
would have been common. The upgrade will still work, just be a bit more
spammy than it used to be.