Reduce the slots melded by beastly appendage (mdonais)
The new melding requirements were picked to disincentivise using BA for
optimal marginal gains long past the point of fun; they are a bit
extreme to the point of disincentivising using BA for anything other
than a starting transmuter who hasn't found aux armour. Reducing the
melding to boot and helmet slots keeps the disincentive for using it on
all characters but makes the spell more flexible and allows it to appeal
to other character types while maintaining a tradeoff of defence for
offence.