Prevent eternal Heaven, again (u/Anno474)
This bug was present in an early version of WJC, fixed in 9dd3d4ce55,
added to the epic bugs list, and then recurred in c9e437d3e5.

Like the original fix, I have added a bunch of future-proofing in case
of this specific bug coming back again, as well as fixup code (instead
of a silent fixup, we will now see a crash triggered in player-reacts,
with fixup on reload). The basic issue is that heavenly storm is
implemented as a dual duration + prop, and keeping those in sync is not
robust. It seems like a better fix, which would take more work, would be
to make the end function a duration end effect, and call all duration
end effects on revival. But I suspect there are non-obvious
complications that explain why revival isn't already handled that way,
so I'm just doing a quick fix for now.