Fix crawlrc-dependent blank screen/crash on turn 0
After b274e4624c7c, we print "You add the spells ... to your library"
even when memorizing from a starting spellbook on turn 0.  This can
cause issues for players who have a force_more_message that matches that
text: for those players, any spellbook background starts with a blank
screen (a promptless force_more), and Earth Elementalists just crash.

Two changes:
* Fix the crash by moving add_held_books_to_library down a few lines, so
it happens after items are linked.
* Change the printed message so that on turn 0 it says "You begin with
the spells" instead of "You add the spells"; that way, it's possible to
target (only) the mid-game version with a force_more_message.