Don't repeat dump_order fields
I can't think of a good use case for repeating dump_order fields, and
allowing it seems to cause more annoyance than anything whenever we
change the defaults, or have an alpha-only default (as is currently the
case with turns_by_level). This version uses the first instance in an
order, so if you want to reorder the defaults it will require removing
and then re-adding.

There's no elegant implementation of arbitrary order + unique elements
in c++ that I know of, and simply keeping two data structures seemed
like the least unaesthetic to me.