Turn +Rats into *Rats
The Ratskin Cloak is yet another of Crawl's many, many very powerful
items that demands considerable micromanagement to unlock its power.
Some items offer a lot of fun when micromanaged, but this big bag o'
rats doesn't quite get there.

Change it from an activated ability to instead spawn rats when the
player is hit. Goal is to have bonus rats as a 'sometimes thing'.
Current odds are somewhat arbitrary but seem to play OK.