Item menu improvements (fixes 12331)
12331 is that (since cc164c81d) a double esc is required to exit from
various item menu prompts, because of a bad continue. When fixing this,
I made several other improvements, namely letting `*` and `?` act as
toggles rather than absolute mode switches. That is, `*` turns on and
off the full inventory display, and `?` turns on and off the menu
altogether, letting you see the map.  This is a behavior change, but I
think it works much better for the menus I tested.  Also, esc for
non-autolist menus does still return you to the prompt, which I think
was previous behavior.  (What will this domino to break that I haven't
tested? tbd. This code is used in a lot of cases.)