Fix draconian acid breath.
It only ever splashed onto monsters that were adjacent to both the spitter
and the target. I believe this was a holdover from when the code was
used for mottled draconian breath, which had a range of 1 anyway. The
targeter always claimed that the splash would hit every monster adjacent
to the target, so let's make that true.

Also stop the breath from splashing the target twice, which doesn't look
intentional either.