Adjust the display of the monster list in webtiles glyphs mode
This change vertically center aligns the health status (colour box) in
the monster list in webtiles glyphs mode and makes it looks more similar
to that of the console version.

The glyphs rendered in the monster list in webtiles glyphs/hybrid modes
were not scaled accordingly to the crawl-internal zooming (rc options or
{} in X), which looks pretty weird and is not consistent with how this
is handled in normal tile mode (i.e., icons rendered in the monster list
are correctly scaled). This change improves the aesthetics and the
consistency between normal tiles mode and glyphs/hybrid mode in
webtiles.