New Misc Item: Condenser Vane
The condenser vane is inteded as a misc-evocable replacement for the
wand of clouds. Wand of clouds used a cone targeter, and targeting cones
really don't fit well with crawl's grid geometry. The new misc evoker
places clouds in a 3x3 arrangement centered on each monster in LOS,
using a similar random pick table to the old wand of clouds, but
permitting high tier clouds all the way from 0 power, just with low
chance, to avoid breakpoints.

This is put on a misc item instead of a wand as a part of larger changes
to evocations with the goal of dividing wands/misc items by
monster-usability. Monsters using wand of clouds tended to only hurt
themselves and not get the player well, and this effect would have
similar issues.