Aesthetic enhancement of webtiles glyph mode
Webtiles glyph mode looked quite different from the console version,
especially with high-DPI scaling enabled. This is because it was using
the font size with some margins added as the cell height; and the glyphs
were vertically aligned to the center of the em-square to prevent them
from being clipped. However, some glyphs will be clipped anyway, as the
font size rarely corresponds to the font height. For example, if a
player is using box-drawing characters such as █ (U+2588), ░ (U+2591), ▒
(U+2592) and ▓ (U+2593) to display various types of walls, the walls
will appear to be unconnected or clipped.

This commit aims to make the dungeon view looks more like how it is
rendered in the console version by following the glyph metrics exactly
as prescribed by the font. With the change, the dungeon view of the
console version and the webtiles glyph mode should be more similar and
the box-drawing characters will no longer appears to be unconnected,
provided that the monospaced font used by the browser is properly
authored in the first place.