Add glyph mode customization options for webtiles
This commit adds two options `glyph_mode_font` and
`glyph_mode_font_size` that allow customization of the font and its size
used to render the dungeon view in the webtiles glyph mode. If set to
monospace (default) or the font is not available, browser's default
monospaced font will be used.