Allow Rampaging on randart armour
Rampaging is currently pretty rare in practice:

* Ego boots are extremely rare spawns at itemgen.
* Many species cannot equip boots.
* Some characters end up with melded boots due to build choices.
* The positioning choices and skilling choices offered by rampaging
  are more impactful when presented earlier in the game.

This commit aims to make rampaging available to a wider range of
characters, by adding it as a potential randart armour property.

ARTP_RAMPAGING generates as 'potentially_good', similar to Harm.

This commit also updates the rampage descriptions slightly,
to remove a minor spoiler for players.