Add Mlioglotl, a new Lugonu-themed unique.
The vision for Mlioglotl is to add a new unique with a strong
flavour themed around Lugonu and the Abyss. In-game description:

"A demon prince once betrayed his realm through a dark ritual to Lugonu.
Becoming the twisted mass of corruption known as Mlioglotl was his
reward.

Seeing such a powerful acolyte of Lugonu within the mortal plane is a
rare sight indeed, and one should fear the denziens and horrors of the
Abyss that are never far behind."

Gameplay flavour vision: A hulking mass ("unique thrashing horror"
wouldn't be far off the mark visually speaking) charging after the
player whilst tearing open rifts between the dungeon and the abyss.

Gameplay design vision: The player is torn between wanting to focus
damage onto Mlioglotl to stop him casting further instances of corruption,
and needing to deal with anything too nasty that is spawned. This is
similar to the dynamics between the player and many of the monsters with
summoning spells in the game, however, as Mlioglotl's corruption is
transferring monsters to the mortal plane rather than summoning them,
monsters that appear as a result of Mlioglotl's corruption are durably
summoned, which adds a relatively novel additional angle to the
interplay between player and 'summoner'. Of course Mlioglotl is no
pushover himself either, and is positioned via depth marker to show up
only later in the game.

The mechanics of Mlioglotl corrupting the dungeon are handled though a
new monster spell 'corrupt_locale'. This effectively triggers a
customised version of the Lugonu corruption player ability with a vastly
toned down duration and toned down monster set.

One concern with any kind of level-altering effect of course would be
that a player could kite the monster around and use them to 'shatter'
open difficult geometry. Mlioglotl is fast, and hits relatively hard,
which perhaps is enough to deter this tedious behaviour, however I'd
welcome feedback on other potential methods to disincentivise this such
as preventing Mlioglotl from using stairs (he has a large tile
currently, so this would fit the visual-language-precedence set by the
Royal Jelly).

Anyway, this is my first time digging into the source code of crawl
beyond vaults and tiles, so apologies in advance for any convention
breaking that I'm not aware of. Always happy to tidy up, make suggested
changes, and generally take my submissions in a direction that others
feel is most healthy for the game.

Thanks,

aMcNicky