Remove staves of energy
Without a hunger mechanic to interact with, staves of energy became a
trade-actions-for-mp-regen item. This is not particularly interesting in
combat as it removes the impact of MP management on a battle and allows
the MP mechanic to be subverted for almost no cost, and full of tedious
incentives out of combat.

If rapid sources of MP are desired, the regen rate of MPRegen amulets
can be increased.