Turn +Fog into *Fog
The cloak of the Thief, like the ratskin cloak, is an overpowered item
to use optimally, providing an effect otherwise locked behind a god
passive or a finite consumable in unlimited quantities for a minimal
tradeoff.

This commit gives the cloak a fixed chance (currently a coinflip) to
generate fog on hit a sufficiently hard hit (> 10% mhp). This is to give
it a different feel from Dithmenos' passive which scales with the hit
magnitude and piety. The sufficiently hard threshold is to prevent
keeping a trivial monster around for free fog on 0 damage hits.