Tweak Maggie description
The old version was a bit plodding. It also didn't even try to
explain how Maggie was apprenticing with the Hell Knights (which
are clearly in the Dungeon) *before* she entered. This version is
playing with hellmonk's parallel-timelines idea by hinting that
Maggie comes from 'a distant time', though I think that's still
probably too subtle...