Give x-v a selection menu when there are many describable things
This commit creates a selection menu when you x-v a tile with multiple
monsters, items, or notable features (following the ^x screen's notion
of notability), whether or not they are of the same type.

As of 2a96157, when there is an item stack on a tile you can x-v the
stash to describe any item in the stash by selecting the desired item
from a menu. However, if a monster is on that stash, then x-v describes
the monster and ignores the items. Similiarly, a monster or items stop
you from seeing the description of an interesting feature. After this
commit a selection menu will pop up in any of those cases, so you won't
be stopped from examining underlying things of a different type.

This commit adapts the menu in _full_describe_menu, which is already
designed for displaying a selection menu with these three types of
things. Therefore I've removed the function describe_items, added
recently by 2a96157 to allow selection from a stack of items.