Following PR#1258, resubmit new arrival vaults
The second PR in the series, submitting amended arrival vaults following
a first round review by ebering. Vaults in this collection:

<simple.des>
amcnicky_arrival_double_pillar
A medium sized arrival vault featuring two highly randomised pillars and
a selection of entrance/exit possibilities

amcnicky_arrival_single_pillar
As above, but on a smaller scale

amcnicky_arrival_feature_with_door
A double-entrance arrival vault with the potential for many different
hints of flavour in the feature tile

amcnicky_arrival_flooded_fountain
The player enters the dungeon through an entry room containing a number
of fountains. The downstairs of an abandonned bath house? A door in the
back of an ancient fountain temple? One of the fountains is broken, and
the player sees a corresponding pool of water

amcnicky_arrival_choice
A vault intended to off the chance for an interesting tactical choice.
Should the player be chased back into this room by an early enemy, there
are no pillars to dance around. Instead, the player must either stand
and fight, or take a risk with the shaft to escape.

amcnicky_arrival_underground_garden
A more decorative vault, the player enters the dungeon by passing
through an underground garden. Was the dungeon entrance hidden within a
forbidden forest?

<small.des>
amcnicky_arrival_blooded_trapdoor
The player is presented with a randomised pillar, two possible
connections to the dungeon, and large smears of blood surrounding a
solitary trapdoor. What became of the last player who tried to use that
trapdoor?

amcnicky_gap_in_the_tavern_cellar
The player enters the dungeon via a secret room behind one of the water
features in an unspecified room. The vault tiles make this connection
through to the dungeon of zot appear almost intentional?

amcnicky_arrival_wind_blow
Which entrance to choose to the great dungeon of Zot? Which way does the
wind blow?

amcnicky_arrival_abandonned_market
The player enters the dungeon of zot through a mysterious back alleyway
amongst an abandonned marketplace.