Add 3 new altar vaults for second round review
With thanks to ebering for reviewing the initial submission, this commit
is the first in a series which resubmits the vaults having been amended
in line with first review commentary. The vaults added are as follows:

<altar.des>
amcnicky_altar_gilded
A simple altar layout with (the potential for) a nice monetary surprise.

amcnicky_altar_inner_sanctum
A less-generic, and more tactically interesting re-think of the vault
previously submitted under the name "amcnicky_altar_medium_temple".
The player is challenged to reach the centre of a 2-ring sanctum, with
the inner ring containing a greater density of monsters as well as the
altar.

amcnicky_altar_lugonu_corruption
Lugonu has entered the dungeon! This vault iterates on the initial
submission idea of 'dungeon corruption' which could generate in many
branches and represents the fist of Lugonu smashing into the terrain.
Current monster spawns use Lurking Horrors to bring their fairly novel
tactical implications beyond just the abyss.

You can find ebering's initial comments as well as my brief working
commentary written in response to the first-round pull request at
github.com/crawl/crawl/pull/1525 though please note that this submission
is in respect of the altar vaults only.

Thanks,

amcnicky