Refactor and generalize quivering
This commit rewrites the quiver to be both more usable and more general,
cleaning up a lot of ncient and hard-to-understand logic.
The basic idea is that what gets quivered is a quiver::action, and this
is stored directly. This class abstracts over many of the details of
what kind of action it is, paving the way for some interesting future
commits. Right now, though, the system just reimplements the old quiver
functionality involving ammo.

I was not able to fully eliminate the old player_quiver, but the
descendent class (quiver::history) is much simpler and used basically
for tracking what was last quivered of any ammo type. Right now, it
still handles serialization as well, and inventory changes. Most of the
weirdest stuff is cleaned up, hopefully.

There may be some logic changes (because old quiver logic was quite
convoluted and I don't necessarily understand everything it did). One
that I do know of is that throwing ammo won't auto-quiver when you run
out of launcher ammo.