Avoid ODR violations in mon-ai-action.h
With -Wmissing-declarations, gcc warns "no previous declaration for"
each of these functions.  Even worse, it violates the one definition
rule to define an extern non-inline function in more than one
translation unit, as would happen if we were to include this header
from anywhere other than mon-cast.cc.

Practically, it would work with most compilers, since they are not
required to flag ODR violations across compilation units, but it would
still be undefined behaviour, and the warnings are annoying anyway.