Reorder the mana link facet.
Provide boosted mp regeneration first, when it's most likely to matter, then give spirit shield second. Mp regen earlier might make it more useful as mp regen rather than as pseudo hp-regen, though overall the change should be and is intended to be a nerf.

The third rank of this facet still leaves something to be desired and has a bad interaction with posion damage (specifically, can kill the player at "safe" levels of poison by redirecting their hp regen), but cannot be moved before spirit shield for obvious reasons. I hesitate to replace it with either additional mp regen or a mana vamp effect because that gives this facet even more overlap with other regen boosting facets. Someone should make gspirit work on poison damage, but that is out of scope for this rework.