Rework Ignite Blood.
The third rank of the ignite blood facet is very cool, but the first two ranks are boring. Split the power of the actual blood igniting part over all three ranks of the facet - an increasing chance to ignite blood and longer duration flame clouds with each rank. Any rank of ignite blood also confers flame cloud immunity, since that's kind of necessary to use the mutation.