Rework Demonic Guardian.
Previously, this facet used an extremely bad tension formula to determine whether or not to summon a guardian. It also had a unique time schedule that practically guaranteed the first rank of the mutation very early and usually pushed the second and third ranks very late. The result was an easily gamed mutation that turbo stomped early dungeon and then kept summoning imps until vaults or later.
This rework ignores tension altogether; instead, the mutation has a chance to activate whenever the player takes damage (guaranteed if the damage is greater than half max hp or if the player's health is critically low) and summons a demon whose tier is partially based on experience level. The variety of potential guardians is expanded a bit as well. Finally, the time schedule is brought into line with every other tier two facet in the game.