Don't crash when putting on non-jewellery
You can try to do this via the * menu, as long as you have jewellery at
all. This looks like it was broken when the call order was rearranged in
b2248e6a10f6; the problem is that jewellery_is_amulet will crash if it
is handed non-jewellery (because it is a subtype check, and is only
meaningful in that case), whereas this code assumes it would return
false. Just adding in the item type check is inelegant, but I don't want
to mess with the ASSERT.