Adjust the icemail facet.
While the icemail itself was very cool and powerful, the previous two ranks of the facet were relatively boring, and freezing cloud immunity was an especially odd choice given that the facet did not create clouds. This rework splits icemail in half across ranks 2 and 3 of the facet and reduces the total AC to 8. A new mutation, Condensation Shield, is added for the first rank, providing 4 SH that dissipates under the same conditions as icemail. This should keep the facet's role relatively intact without buffing it too much lategame; total defenses provided are 4 SH and 8 AC at rank 3.
Because the monster enchantment for icemail also uses ICEMAIL_MAX, I buffed gelid demonspawn base AC by 2 to compensate for their reduced icemail AC.