Clean up evoke checks a bit
Move more checks into evoke_check, and use it more widely (including for
quivering wands).

This handles zigfigs and reaching slightly differently than in the past;
while berserk (etc) V is now possible if either of these are in play,
but will only show the currently enabled items in the menu.