Properly handle elemental and cloud producing facets.
These mutations are meant to exclude each other so that the player gets at most one elemental and at most one cloud facet. Formerly, this took advantage of mutations with a single mutlevel to handle the check, but that doesn't work anymore. This is as quick and dirty a fix as I could manage, since I plan on looking at this again for the tier 3 mut pr.