Re-implement tossing
Tossing was temporarily impossible after throwing get refactored
entirely into the fire system. This adds an action type for tossing that
is not used unless items are force-quivered. It is currently not
possible to toss items that can be otherwise quivered, including wands.

This allows quivering items for tossing. There was some discussion
in ##crawl-dev on whether this was a good idea, and I'm not sure it is;
but as long as there is tossing at all, there is no reason to make it
impossible to do via the quiver (and in fact with the current API, that
would be more work rather than less). I don't particularly want to take
a stand on whether tossing is a good idea in the first place as part of
this commit series, so I leave things as they were.