Raise Zot Clock bank cap 15k -> 27k
The goal of the zot clock bank cap is to prevent players from having
truly tremendous amounts of time to spend to do tedious things on
specific floors. In practice, I haven't seen reports of anyone really
doing anything unpleasant with the current limit, and the 15k cap caused
issues with shaft traps on rare occasions.

Let's raise it to 27k (meaning it approximately doesn't exist outside D)
and see what happens. It might vanish entirely in future, or it might drop
again, depending on the results of this experiment.