Remove some boring beasts
With the re-introduction of boulder beetles to Lair and Dungeon, it's a
good time to remove some of the redundant beasts we have in these
branches. This commit removes hippogriffs, crocodiles, and porcupines.
The first two are simple speed 10 beasts that have little to
differentiate themselves from other melee-only monsters of similar
difficultly at mid-Dungeon and Lair depths.

Porcupines are fast and have a spine mechanic, so they're arguably the
most interesting of the three removed monsters. However their spines are
an "anti-melee" trait that works poorly in terms of influencing the
player's decision to melee or not. Culling the monster list down to a
set that all work well design-wise is a good long-term goal, and neither
branch lacks for good fast melee monsters.

Level spawn weights of each of the three removed monsters has been
giving to respective similar monsters in both branches, with adjustments
so that the impact of removal is minimal. The new weights and changes in
level spawn chance are described in this google spreadsheet:

https://docs.google.com/spreadsheets/d/16lBJeSFqxFsy8MPFUBcjV0ZnvU172e2G-8WWO1lyxpU/edit?usp=sharing

See the "Chance (Modified)" and "Increase in chance" sections of each
sheet.

Dungeon has such a large monster set that the largest increase in any
affected monster placement on any level is a fraction of 1%. For Lair,
the largest increase is less than 3%. I increased placement range of the
few more interesting beasts such as polar bears, hell rats, stream
dragons. Additionally, a few weights for rarer Lair beasts are now
doubled, but these monsters are already quite rare and have a placement
chance of 1% or less throughout their range.