Remove unused Pakellas vaults
These vaults where sitting in a des file with 0 weight mostly due to
accompanying lua triggers that need to be defined for save compatibilty.
I've moved the required functions to a designated save compat lua file
with the function definitions emptied out to the extent possible. This
should cover save compatibilty, but even if there is a problem, the odds
of there being an affected save out there with the vaults actually
generated are quite small at this point.